+++ lib/oneview-sdk/resource/api300/c7000/logical_interconnect.rb
@@ -10,12 +10,14 @@
 # language governing permissions and limitations under the License.
 
 require_relative '../../api200/logical_interconnect'
+require_relative 'scope'
 
 module OneviewSDK
   module API300
     module C7000
       # Logical interconnect resource implementation for API300 C7000
       class LogicalInterconnect < OneviewSDK::API200::LogicalInterconnect
+        include OneviewSDK::API300::C7000::Scope::ScopeHelperMethods
 
         # Create a resource object, associate it with a client, and set its properties.
         # @param [OneviewSDK::Client] client The client object for the OneView appliance
@@ -26,26 +28,6 @@ module OneviewSDK
           # Default values:
           @data['type'] ||= 'logical-interconnectV300'
           super
-        end
-
-        # Lists internal networks on the logical interconnect
-        # @return [OneviewSDK::Resource] List of networks
-        def list_vlan_networks
-          ensure_client && ensure_uri
-          results = OneviewSDK::Resource.find_by(@client, {}, @data['uri'] + '/internalVlans')
-          internal_networks = []
-          results.each do |vlan|
-            net = if vlan['generalNetworkUri'].include? 'ethernet-network'
-                    OneviewSDK::API300::C7000::EthernetNetwork.new(@client, uri: vlan['generalNetworkUri'])
-                  elsif vlan['generalNetworkUri'].include? 'fc-network'
-                    OneviewSDK::API300::C7000::FCNetwork.new(@client, uri: vlan['generalNetworkUri'])
-                  else
-                    OneviewSDK::API300::C7000::FCoENetwork.new(@client, uri: vlan['generalNetworkUri'])
-                  end
-            net.retrieve!
-            internal_networks.push(net)
-          end
-          internal_networks
         end
 
         # Updates settings of the logical interconnect