+++ spec/integration/resource/api200/enclosure_group/update_spec.rb
@@ -1,19 +1,7 @@
 require 'spec_helper'
 
 RSpec.describe OneviewSDK::EnclosureGroup, integration: true, type: UPDATE do
-  include_context 'integration context'
-
-  describe '#script' do
-    it 'can retrieve the script' do
-      item = OneviewSDK::EnclosureGroup.find_by($client, name: ENC_GROUP_NAME).first
-      script = item.get_script
-      expect(script).to be_a(String)
-    end
-
-    it 'can set the script' do
-      item = OneviewSDK::EnclosureGroup.find_by($client, name: ENC_GROUP_NAME).first
-      item.set_script('#TEST COMMAND')
-      expect(item.get_script.tr('"', '')).to eq('#TEST COMMAND')
-    end
-  end
+  let(:current_client) { $client }
+  include_examples 'EnclosureGroupUpdateExample', 'integration context'
+  include_examples 'EnclGroupScriptC7000Example', 'integration context'
 end