+++ spec/integration/resource/api500/c7000/logical_interconnect/create_spec.rb
@@ -1,0 +1,6 @@
+require 'spec_helper'
+
+klass = OneviewSDK::API500::C7000::LogicalInterconnect
+RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
+  # Cannot create individually
+end